+++ src/tests/integration/steps/run-background.ts
@@ -1,10 +1,10 @@
 // tslint:disable:no-let
 import { expect } from 'chai'
 import { Map } from 'immutable'
-import { Mocks } from './mocks'
-import { run } from '../background/run'
-import { ensureActiveTab } from '../selectors'
-import { getPlainText } from '../draft-js-utils'
+import { Mocks } from '../mocks'
+import { run } from '../../../background/run'
+import { ensureActiveTab } from '../../../selectors'
+import { getPlainText } from '../../../draft-js-utils'
 
 export function runBackground(mocks: Mocks, opts: { allowActionFailure: boolean } = { allowActionFailure: false }) {
   let unsubscribe
